nyc_data_producer/requirements.txt
azure-storage-blob
azure-eventhub